Rename project to fact sorter
