+++ README.md
@@ -1,3 +1,3 @@
-# Eventbus Learning
+# Fact Sorter
 
 A project to help me better understand how AWS Eventbuses work including how create one and connect other services to it.
