+++ pytest.ini
@@ -1,7 +1,7 @@
 [pytest]
 pythonpath =
-    eventbus_learning/
-    eventbus_learning/application
+    fact_sorter/
+    fact_sorter/application
     cdk/
 env =
     AWS_ACCOUNT_ID=1234567890
